requirements.txt
azure-cognitiveservices-vision-customvision==1.0.0
Pillow==6.0.0
Flask==0.12.2
Flask-API==1.0
